bootstrap: freebsd: run chflags -R noschg on fresh chroot
When rootfs tarball is unpacked, some files are flagged as schg
(system immutable), so even root can't remove the chroot tree entirely.

Since our chroot's are just temporary playgrounds (and not security-related
at all), expected to be thrown away anytime, this makes it easier for
the user to remove them.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>